+++ routers/__init__.py
@@ -7,6 +7,7 @@ from .sharepoint import router as sharepoint_router
 from .sources import router as sources_router
 from .directories import router as directories_router
 from .schedules import router as schedules_router
+from .nlp import router as nlp_router
 
 __all__ = [
     'sharepoint_router',
@@ -16,4 +17,5 @@ __all__ = [
     'sources_router',
     'directories_router',
     'schedules_router',
+    'nlp_router',
 ]